MODPATBLK-41 refactoring of GenerationService
